homework3/requirements.txt
bentoml==1.4.25
torch
pillow
transformers
requests